improve code and support adding `csp nonce`
